hp-6: WIP: So much broken!
