+++ src/hyper-model/hypermodel/__init__.py
@@ -1,7 +1,5 @@
 from hypermodel.utilities.hm_shell import sh
 from hypermodel.kubeflow.deploy_dev import deploy_to_dev
 
-from hypermodel.kubeflow.decorators import op
-
 import logging
 import sys
