+++ src/hyper-model/hypermodel/features/categorical.py
